backend/requirements.txt
flask
flask-cors
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
requests
